move and update /contact page
